Split up DataRepo/tests/test_formats.py into test files under DataRepo/tests/formats and tweaked the test stub generation code.
Files checked in: DataRepo/formats/dataformat_group.py DataRepo/tests/test_formats.py DataRepo/tests/tracebase_test_case.py DataRepo/tests/formats/formats_test_base.py DataRepo/tests/formats/test_dataformat.py DataRepo/tests/formats/test_dataformat_group.py DataRepo/tests/formats/test_dataformat_group_query.py DataRepo/tests/formats/test_fluxcirc_dataformat.py DataRepo/tests/formats/test_peakdata_dataformat.py DataRepo/tests/formats/test_peakgroups_dataformat.py
